better handling of {} surrounded authors
This commit resolves an issues with split_names() that it will not
handle `{author}` entries and drop these. Following through git history,
it is not clear why this is. My only guess is that this is to ensure
that functions like `splitFullName()` don't error out. So for Bibtex entry:

```bibtex
@article{test01,
  author="{The Company Inc. & Co.}",
  ...
}
```

calling `getRawAuthors()` returns an empty array!

TBC anon.